telegram-bot/requirements.txt
python-telegram-bot==21.10
requests==2.32.3
openai==1.53.0
fastapi==0.115.8
envs==1.4
httpx==0.27.2
